# src/swagger/paymentFrequenciesSwagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Frecuencias de Pago",
    "version": "1.0.0"
  },
  "paths": {
    "/payment_frequencies": {
      "post": {
        "summary": "Crear una nueva frecuencia de pago",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentFrequency"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Frecuencia de pago creada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentFrequency"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Obtener todas las frecuencias de pago",
        "responses": {
          "200": {
            "description": "Lista de frecuencias de pago",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PaymentFrequency"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/payment_frequencies/{id}": {
      "get": {
        "summary": "Obtener una frecuencia de pago por ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Frecuencia de pago",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentFrequency"
                }
              }
            }
          },
          "404": {
            "description": "Frecuencia de pago no encontrada"
          }
        }
      },
      "put": {
        "summary": "Actualizar una frecuencia de pago",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentFrequency"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Frecuencia de pago actualizada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentFrequency"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Eliminar una frecuencia de pago",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Frecuencia de pago eliminada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentFrequency"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "PaymentFrequency": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "ID de la frecuencia de pago"
          },
          "name": {
            "type": "string",
            "description": "Nombre de la frecuencia de pago"
          },
          "description": {
            "type": "string",
            "description": "Descripción de la frecuencia de pago"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Fecha de creación de la frecuencia de pago"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Fecha de actualización de la frecuencia de pago"
          }
        }
      }
    }
  }
}